// types/my-reservations.ts
export interface Reservations {
  id: number;
  status: string;
  totalPrice: number;
  headCount: number;
  nickname: string;
  userId: number;
  date: string;
  startTime: string;
  endTime: string;
  createdAt: string;
  updatedAt: string;
  activityId: number;
  scheduleId: number;
  reviewSubmitted: boolean;
  teamId: string;
}

export interface ReservationsResponse {
  reservations: Reservations[];
  totalCount: number;
  cursorId: number;
}